Add abstract entity service
Added abstract service that that implements the following operations with an entity: getting all records, getting a specific object by id, adding a new one and deleting by id. Each class that extends it must specify the entity type in the parameter and set the entity class via setter in the constructor.